+++ .yeoman/index.js
@@ -18,7 +18,7 @@ module.exports = class extends Generator {
     const props = await this.prompt(questions);
     this.props = props;
     this.props.prettyName = prettyName(this.props.name);
-    this.props.componentName = prettyName(this.props.name, '');
+    this.props.packageName = prettyName(this.props.name, '');
     this.props.dependencies = {
       ...basePackage.dependencies,
       ...basePackage.devDependencies,
@@ -28,16 +28,16 @@ module.exports = class extends Generator {
   paths() {
     this.log(chalk.blue('Creating your package folder...'));
 
-    const destination = `./packages/components/psammead-${this.props.name}`;
+    const destination = `./packages/${this.props.kind}/psammead-${this.props.name}`;
     shell.mkdir('-p', destination);
     this.destinationRoot(destination);
-    this.sourceRoot(`${__dirname}/templates`);
+    this.sourceRoot(`${__dirname}/templates/${this.props.kind}`);
   }
 
   writing() {
     this.log(chalk.blue('Generating your package files...'));
 
-    fs.readdir(`${__dirname}/templates`, (err, files) => {
+    fs.readdir(`${__dirname}/templates/${this.props.kind}`, (err, files) => {
       files.forEach(filename => {
         this.fs.copyTpl(
           this.templatePath(filename),
@@ -57,7 +57,7 @@ module.exports = class extends Generator {
   end() {
     this.log(
       chalk.green(
-        `Success! You can now start developing on your new Psammead component.\nYour new package can be found here: ./packages/components/psammead-${this.props.name}`,
+        `Success! You can now start developing on your new Psammead package.\nYour new package can be found here: ./packages/${this.props.kind}/psammead-${this.props.name}`,
       ),
     );
   }