+++ packages/utilities/psammead-storybook-helpers/src/withServicesKnob.test.jsx
@@ -2,7 +2,7 @@ import { render, waitForDomChange } from '@testing-library/react';
 import * as knobs from '@storybook/addon-knobs';
 import * as scripts from '@bbc/gel-foundations/scripts';
 import withServicesKnob from './withServicesKnob';
-import LANGUAGE_VARIANTS from './text-variants';
+import TEXT_EXAMPLES from './text-variants';
 
 it('should correctly configure the default story book dropdown', () => {
   const storyFn = () => {};
@@ -12,7 +12,7 @@ it('should correctly configure the default story book dropdown', () => {
 
   expect(knobs.select).toHaveBeenCalledWith(
     'Select a service',
-    Object.keys(LANGUAGE_VARIANTS),
+    Object.keys(TEXT_EXAMPLES),
     'news',
   );
 });
@@ -63,8 +63,11 @@ it('should pass the correct props to the story function', () => {
     script: scripts.latin,
     service: 'news',
     text: 'Could a computer ever create better art than a human?',
+    longText:
+      'The critic, author, poet and TV host was known for his witty commentary on international television.',
     dir: 'ltr',
     locale: 'en',
+    variant: 'default',
   };
 
   expect(mockStoryFn).toHaveBeenCalledWith(expected);
@@ -79,10 +82,32 @@ it('should pass the correct chosen service props to the story function', () => {
   const expected = {
     script: scripts.arabic,
     service: 'arabic',
-    text:
-      'المحكمة العليا الأمريكية ذات الأغلبية المحافظة وافقت على احتجاز غير المواطنين لأجل غير مسمى حتى بعد سنوات من خروجهم من السجن',
+    text: 'لماذا يخجل البعض من اسم قريته في مصر؟',
+    longText:
+      'هناك وقائع عدة تتسم بالسخرية والجدل والتنمر، ضد أهل القرية الذين أصابهم الغضب والسخط مما دفعهم إلى تقديم طلب لتغيير اسم قريتهم.',
     dir: 'rtl',
     locale: 'ar',
+    variant: 'default',
+  };
+
+  expect(mockStoryFn).toHaveBeenCalledWith(expected);
+});
+
+it('should pass the correct chosen service props to the story function', () => {
+  const mockStoryFn = jest.fn();
+  knobs.select = () => 'ukChinaSimp';
+
+  withServicesKnob()(mockStoryFn);
+
+  const expected = {
+    script: scripts.chinese,
+    service: 'ukchina',
+    text: '该计划的批评者说，这个政策不能解决住房短缺的问题',
+    longText:
+      '但在当今世界，尽管许多人已不再把步行作为一种主要的出行方式，但巴黎仍然是属于孤僻、哲学式观察者的理想城市。毕竟，法国人习惯于花时间以文学和哲学的方式观察和思考周围的环境',
+    dir: 'ltr',
+    locale: 'zh-cn',
+    variant: 'simp',
   };
 
   expect(mockStoryFn).toHaveBeenCalledWith(expected);