+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 2.0.23 | [PR#2714](https://github.com/bbc/psammead/pull/2714) Talos - Bump Dependencies - @bbc/psammead-storybook-helpers |
+| 2.0.22 | [PR#2611](https://github.com/bbc/psammead/pull/2611) Improve yeoman generator, internal refactor to introduce seperate package templates |
+| 2.0.21 | [PR#2708](https://github.com/bbc/psammead/pull/2708) Talos - Bump Dependencies - @bbc/psammead-image-placeholder, @bbc/psammead-media-indicator |
+| 2.0.20 | [PR#2705](https://github.com/bbc/psammead/pull/2705) Talos - Bump Dependencies - @bbc/psammead-assets |
+| 2.0.19 | [PR#2703](https://github.com/bbc/psammead/pull/2703) Talos - Bump Dependencies - @bbc/psammead-caption, @bbc/psammead-copyright, @bbc/psammead-image-placeholder, @bbc/psammead-inline-link, @bbc/psammead-media-indicator, @bbc/psammead-paragraph, @bbc/psammead-timestamp |
+| 2.0.18 | [PR#2701](https://github.com/bbc/psammead/pull/2701) Talos - Bump Dependencies - @bbc/psammead-assets, @bbc/psammead-styles |
+| 2.0.17 | [PR#2700](https://github.com/bbc/psammead/pull/2700) Talos - Bump Dependencies - @bbc/psammead-caption, @bbc/psammead-copyright, @bbc/psammead-image-placeholder, @bbc/psammead-inline-link, @bbc/psammead-media-indicator, @bbc/psammead-paragraph, @bbc/psammead-timestamp |
+| 2.0.16 | [PR#2577](https://github.com/bbc/psammead/pull/2577) Add `react-test-renderer` and `@testing-library/react-hooks` |
+| 2.0.15 | [PR#2697](https://github.com/bbc/psammead/pull/2697) Talos - Bump Dependencies - @bbc/psammead-storybook-helpers, @bbc/psammead-styles |
+| 2.0.14 | [PR#2615](https://github.com/bbc/psammead/pull/2611) Improve yeoman generator, to specify template types |
+| 2.0.13 | [PR#2687](https://github.com/bbc/psammead/pull/2687) Talos - Bump Dependencies - @bbc/psammead-storybook-helpers |
+| 2.0.12 | [PR#2666](https://github.com/bbc/psammead/pull/2666) Install Ramda as a dev dependencies at the root package.json |
+| 2.0.11 | [PR#2681](https://github.com/bbc/psammead/pull/2681) Talos - Bump Dependencies - @bbc/psammead-storybook-helpers |
 | 2.0.10 | [PR#2666](https://github.com/bbc/psammead/pull/2666) Talos - Bump Dependencies - @bbc/psammead-image-placeholder, @bbc/psammead-media-indicator |
 | 2.0.9 | [PR#2657](https://github.com/bbc/psammead/pull/2657) Update codeowners |
 | 2.0.8 | [PR#2663](https://github.com/bbc/psammead/pull/2663) Talos - Bump Dependencies - @bbc/psammead-assets |