+++ packages/components/psammead-media-player/CHANGELOG.md
@@ -3,6 +3,12 @@
 <!-- prettier-ignore -->
 | Version       | Description                                                                                                                  |
 | ------------- | ---------------------------------------------------------------------------------------------------------------------------- |
+| 2.3.8 | [PR#2708](https://github.com/bbc/psammead/pull/2708) Talos - Bump Dependencies - @bbc/psammead-play-button |
+| 2.3.7 | [PR#2705](https://github.com/bbc/psammead/pull/2705) Talos - Bump Dependencies - @bbc/psammead-assets |
+| 2.3.6 | [PR#2703](https://github.com/bbc/psammead/pull/2703) Talos - Bump Dependencies - @bbc/psammead-play-button |
+| 2.3.5 | [PR#2701](https://github.com/bbc/psammead/pull/2701) Talos - Bump Dependencies - @bbc/psammead-assets |
+| 2.3.4 | [PR#2700](https://github.com/bbc/psammead/pull/2700) Talos - Bump Dependencies - @bbc/psammead-play-button |
+| 2.3.3 | [PR#2684](https://github.com/bbc/psammead/pull/2684) Added scrolling attribute to disable scrolling on canonical and amp iframe |
 | 2.3.2 | [PR#2666](https://github.com/bbc/psammead/pull/2666) Talos - Bump Dependencies - @bbc/psammead-play-button |
 | 2.3.1 | [PR#2663](https://github.com/bbc/psammead/pull/2663) Talos - Bump Dependencies - @bbc/psammead-assets |
 | 2.3.0 | [PR#2628](https://github.com/bbc/psammead/pull/2628) Add guidance message for assistive technology in Play Button and hide it in Guidance |