+++ packages/utilities/psammead-styles/CHANGELOG.md
@@ -3,6 +3,8 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 4.1.2 | [PR#2686](https://github.com/bbc/psammead/pull/2686) Add colour C_ARCHIVE_BLUE and set font families to Latin Reith for Archive |
+| 4.1.1 | [PR#2661](https://github.com/bbc/psammead/pull/2661) Remove font 'Noto Sans Devanagari' from Hindi & Nepali font-family and add bold weight for Telugu service's fonts. |
 | 4.1.0 | [PR#2514](https://github.com/bbc/psammead/pull/2514) Add SerifLight and SansLight functions |
 | 4.0.4 | [PR#2471](https://github.com/bbc/psammead/pull/2471) Switch Russian service to use Reith fonts |
 | 4.0.3 | [PR#2404](https://github.com/bbc/psammead/pull/2404) replace inputProvider and dirDecorator with withServicesInput |